Tidied up the codebase a little
Moved to a revealing prototype pattern and combined functions to Ultrawide.apply. Wrapped files in IIFEs for isolation.

Not entirely sure what the 100ms setTimeout was about, can be reinstated if necessary.
Similarly, not sure why classCheck wasn't called in some scenarios when it seemed to be necessary.

Cached DOM selectors during calculations.

The modes could be captured in a Strategy pattern to reduce the need for case statements and boilerplate in future.